backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
numpy
faiss-cpu
sentence-transformers
torch
transformers
pdfplumber
